Add produce_block_with_extrinsics to MockPrimaryNode
produce_block_with_extrinsics is useful when we want to control the exact extrinsic
included in block or for extrinsic that may fail on verification but we still want
to include it in block.

This commit also extract the new block logging in a separated function
so both produce_block_with_slot and produce_block_with_extrinsics can reuse it.

Signed-off-by: linning <linningde25@gmail.com>